refactor: share process_dataframe
